Minor code cleanups.  The most significant is letting the face indices go past 65536 by changing the data type from unsigned short to an Integer.
git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@11517 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 608a8e56c3a1ad8186ed6460522f2d1642c3a0b1

+++ src/gui/rendering/ModelObject.hpp
@@ -76,7 +76,10 @@ typedef struct
    char name[255];                     // The name of the material
    char texture_name[255];             // The file name of the texture associated with the material
    int num_faces;                      // The number of faces that use this material
-   unsigned short faces[MAX_POLYGONS]; // The indices of the faces using this material
+
+   // This needs to be an int since a model can have more than 65536 faces; not making it unsigned for now
+//   unsigned short faces[MAX_POLYGONS]; // The indices of the faces using this material
+   int faces[MAX_POLYGONS];            // The indices of the faces using this material
    rgba_type mat_ambient;              // The ambient color of the material
    rgba_type mat_diffuse;              // The diffuse color of the material
    rgba_type mat_specular;             // The specular color of the material
